refactor: rewire main routing (Phase 6)
